Convert ibazel to library, remove profiler, add listeners
Remove the main*.go files from ibazel and update the package from
main to ibazel so it may be used as a library. Remove profiler
and instead take in a list of listeners to IBazel.New for lifecycle
events. Combine in a single commit for easier rebasing in the future